+++ graph_games/pong.c
@@ -49,7 +49,7 @@ int         bsx;          /* ball position save x */
 int         bsy;          /* ball position save y */
 int         baltim;       /* ball start timer */
 ami_evtrec   er;           /* event record */
-int         jchr;         /* number of pixels to joystick movement */
+long        jchr;         /* number of pixels to joystick movement */
 int         score;        /* score */
 int         scrsiz;       /* score size */
 int         scrchg;       /* score has changed */
@@ -282,7 +282,7 @@ int main(void)
     failtimer = 0; /* clear fail timer */
     ami_opensynthout(AMI_SYNTH_OUT); /* open synthesizer */
     ami_instchange(AMI_SYNTH_OUT, 0, 1, AMI_INST_LEAD_1_SQUARE);
-    jchr = INT_MAX/((ami_maxxg(stdout)-2)/2); /* find basic joystick increment */
+    jchr = LONG_MAX/((ami_maxxg(stdout)-2)/2); /* find basic joystick increment */
     ami_curvis(stdout, FALSE); /* remove drawing cursor */
     ami_auto(stdout, FALSE); /* turn off scrolling */
     ami_font(stdout, AMI_FONT_SIGN); /* sign font */
@@ -363,7 +363,7 @@ int main(void)
 
                   	nottim--; /* derement */
                    	if (nottim == 0) /* times up, turn note off */
-                      	ami_noteoff(AMI_SYNTH_OUT, 0, 1, WALLNOTE, INT_MAX);
+                      	ami_noteoff(AMI_SYNTH_OUT, 0, 1, WALLNOTE, LONG_MAX);
 
                 }
                 /* if the fail note timer is running, decrement it */
@@ -371,7 +371,7 @@ int main(void)
 
                    	failtimer = failtimer-1; /* derement */
                    	if (!failtimer) /* times up, turn note off */
-                      	ami_noteoff(AMI_SYNTH_OUT, 0, 1, FAILNOTE, INT_MAX);
+                      	ami_noteoff(AMI_SYNTH_OUT, 0, 1, FAILNOTE, LONG_MAX);
 
                 }
                 if (ball.x1 > 0) { /* ball on screen */
@@ -386,7 +386,7 @@ int main(void)
                       	bdx = -bdx; /* change direction */
                       	offrect(&ball, bdx, bdy); /* recalculate */
                       	/* start bounce note */
-                      	ami_noteon(AMI_SYNTH_OUT, 0, 1, WALLNOTE, INT_MAX);
+                      	ami_noteon(AMI_SYNTH_OUT, 0, 1, WALLNOTE, LONG_MAX);
                       	nottim = BNCENOTE; /* set timer */
 
                    } else if (intrect(&ball, &wallt)) { /* hits top */
@@ -395,7 +395,7 @@ int main(void)
                       	bdy = -bdy; /* change direction */
                      	offrect(&ball, bdx, bdy); /* recalculate */
                       	/* start bounce note */
-                      	ami_noteon(AMI_SYNTH_OUT, 0, 1, WALLNOTE, INT_MAX);
+                      	ami_noteon(AMI_SYNTH_OUT, 0, 1, WALLNOTE, LONG_MAX);
                       	nottim = BNCENOTE; /* set timer */
 
                    } else if (intrect(&ball, &paddle)) {
@@ -411,7 +411,7 @@ int main(void)
                       	score = score+1; /* count hits */
                       	scrchg = TRUE; /* set changed */
                       	/* start bounce note */
-                      	ami_noteon(AMI_SYNTH_OUT, 0, 1, WALLNOTE, INT_MAX);
+                      	ami_noteon(AMI_SYNTH_OUT, 0, 1, WALLNOTE, LONG_MAX);
                       	nottim = BNCENOTE; /* set timer */
 
                    }
@@ -421,7 +421,7 @@ int main(void)
                         clrrect(&ball); /* set ball not on screen */
                         baltim = NEWBAL; /* start time on new ball wait */
                         /* start fail note */
-                        ami_noteon(AMI_SYNTH_OUT, 0, 1, FAILNOTE, INT_MAX);
+                        ami_noteon(AMI_SYNTH_OUT, 0, 1, FAILNOTE, LONG_MAX);
                         failtimer = FAILTIME; /* set timer */
 
                    } else { /* ball in play */
